requirements.txt
anyio==4.3.0
asyncpg==0.29.0
bcrypt==4.1.2
black==24.2.0
cryptography==42.0.5
fastapi==0.110.0
httpx==0.27.0
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.6.2
pydantic-settings==2.2.1
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-jose==3.3.0
SQLAlchemy==2.0.27
uvicorn==0.27.1
yoyo-migrations==8.2.0